feat: Improve release, testing and launchd daemon reliability
This commit introduces a number of improvements to the project,
including:

- GitHub Actions: The release workflow now builds and pushes
multi-arch Docker images for linux/amd64 and linux/arm64.

- Makefile: The default target is now install. The install.launchd
and uninstall.launchd targets now handle the socket_vmnet_wrapper.sh
script. The uninstall-pvmlab target now runs pvmlab clean --purge
before removing the binary.

- launchd: The socket_vmnet launchd daemon is now more reliable.
It now uses a wrapper script to generate a persistent network
identifier for socket_vmnet. This ensures that the same network
identifier is used across restarts of the service. The launchd
daemon is now loaded and unloaded with launchctl load -w and
launchctl unload -w respectively. This ensures that the daemon
is automatically started at boot and that the service is correctly
unloaded.

- Integration Tests: The integration tests have been refactored
to be more robust and easier to maintain. The socket_vmnet setup
has been moved to a separate file. The tests now include more
checks to ensure that the VMs are correctly configured.

- README: The README has been updated with more examples and
clarifications.
