+++ .github/workflows/pxeboot-container-release.yml
@@ -1,4 +1,4 @@
-name: Release
+name: Pxeboot Docker Container Release
 
 on:
   push:
@@ -23,9 +23,16 @@ jobs:
           username: ${{ github.actor }}
           password: ${{ secrets.GITHUB_TOKEN }}
 
-      - name: Build and push Docker image
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v3
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v3
+
+      - name: Build and push Docker image (x86_64 and aarch64)
         uses: docker/build-push-action@v5
         with:
+          platforms: linux/amd64,linux/arm64
           context: ./pxeboot_stack
           push: true
           tags: ghcr.io/${{ github.repository }}/pxeboot_stack:${{ github.ref_name }}
